[feat] AUTH et DB LIMTED ACCESS à la DB par USER_ID OK

+++ client/src/pages/Ma_bibliotheque.tsx
@@ -1,16 +1,19 @@
 import { useEffect, useState } from "react";
 import { slide as Menu } from "react-burger-menu";
 import { Link, useNavigate } from "react-router-dom";
+
 import "../styles/Book.css";
 import "../styles/BurgerMenu.css";
 import "../styles/Buttons.css";
+
 import AddBook from "../components/AddBook";
 import AddBookManually from "../components/AddBookManually";
 import Book from "../components/Book";
 import DeleteConfirmationModale from "../components/DeleteConfirmationModale";
 import Header from "../components/Header";
 import SearchBar from "../components/Searchbar";
-import AddBook from "../components/AddBook";
+import Cookies from "js-cookie";
+import { useAuth } from "../context/AuthContext";
 
 interface BookProps {
   titre: string;
@@ -24,6 +27,7 @@ interface BookProps {
 }
 
 function Ma_bibliotheque() {
+  const { userId, setUserId } = useAuth();
   const navigate = useNavigate();
   const [books, setBooks] = useState<BookProps[]>([]);
   const [menuOpen, setMenuOpen] = useState<boolean>(false);
@@ -43,10 +47,13 @@ function Ma_bibliotheque() {
 
   /*Récupération des livres*/
   useEffect(() => {
+    if (!userId) {
+      return setUserId(Number(Cookies.get("user_id")));
+    }
     const fetchBooks = async () => {
       try {
         const response = await fetch(
-          "http://localhost:3310/api/livres_with_exemplaires",
+          `http://localhost:3310/api/${userId}/livres_with_exemplaires`,
         );
         const data = await response.json();
         setBooks(data);
@@ -57,7 +64,7 @@ function Ma_bibliotheque() {
     };
 
     fetchBooks();
-  }, []);
+  }, [userId, setUserId]);
 
   /*Gère l'ouverture et la fermeture du menu tiroir*/
   const handleMenuStateChange = (state: { isOpen: boolean }) => {
@@ -131,11 +138,14 @@ function Ma_bibliotheque() {
 
   /* Fonction pour confirmer la suppression */
   const handleConfirmDelete = async () => {
+    if (!userId) {
+      return setUserId(Number(Cookies.get("user_id")));
+    }
     if (!bookToDelete) return;
 
     try {
       const response = await fetch(
-        `http://localhost:3310/api/livres/${bookToDelete.ISBN}`,
+        `http://localhost:3310/api/${userId}/livres/${bookToDelete.ISBN}`,
         {
           method: "DELETE",
         },
